# Backend/api/laravel-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto Final Grado API",
        "description": "Documentacion de la API para el Proyecto Final de Grado",
        "contact": {
            "name": "Alejandro Solar, David Kovacs",
            "email": "alejandrosolarruiz@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/juego-imagenes": {
            "get": {
                "tags": [
                    "JuegoImagen"
                ],
                "summary": "Listar todas las imágenes de los juegos",
                "operationId": "9cd58d5774c04ccd0da050d35851b73f",
                "responses": {
                    "200": {
                        "description": "Lista de imágenes de los juegos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/JuegoImagen"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "JuegoImagen"
                ],
                "summary": "Crear una nueva imagen para un juego",
                "operationId": "059f9045dea941d3a7ce5a985ed02a6f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JuegoImagenCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Imagen creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JuegoImagen"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/juego-imagenes/{id}": {
            "get": {
                "tags": [
                    "JuegoImagen"
                ],
                "summary": "Obtener una imagen por ID",
                "operationId": "a54a08836f221451d445358e937633a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la imagen",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Imagen encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JuegoImagen"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Imagen no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "JuegoImagen"
                ],
                "summary": "Actualizar una imagen existente",
                "operationId": "6ad81f96a62d76bb69f8f73a0b9d8cf2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la imagen",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JuegoImagenUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Imagen actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JuegoImagen"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Imagen no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "JuegoImagen"
                ],
                "summary": "Eliminar una imagen de un juego",
                "operationId": "dbc5e0200f92268ecefd2eac5cfd8b41",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la imagen",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Imagen eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Imagen eliminada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Imagen no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "JuegoImagen": {
                "properties": {
                    "id_imagen": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_juego": {
                        "type": "integer",
                        "example": 2
                    },
                    "url": {
                        "type": "string",
                        "example": "https://example.com/imagen.jpg"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Imagen de portada del juego",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "JuegoImagenCreate": {
                "required": [
                    "id_juego",
                    "url"
                ],
                "properties": {
                    "id_juego": {
                        "type": "integer",
                        "example": 2
                    },
                    "url": {
                        "type": "string",
                        "example": "https://example.com/imagen.jpg"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Imagen de portada del juego",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "JuegoImagenUpdate": {
                "properties": {
                    "url": {
                        "type": "string",
                        "example": "https://example.com/imagen.jpg"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Imagen de portada del juego",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}